+++ CHANGELOG.md
@@ -58,6 +58,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed all scripts in the example folder.
 - Fixed `file_get_html` to return false if the file size is larger than `maxLen`.
 - Fixed a bug that caused the parser to convert UTF-8 to UTF-8 on mistake.
+- Fixed `simple_html_dom::loadFile` to properly forward arguments to `simple_html_dom::load_file`.
 ### Security
 
 ## [1.9] - 2019-05-30